requirements.txt
asgiref==3.7.2
Django==4.2.10
django-js-asset==2.2.0
django-mptt==0.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
